+++ index.html
@@ -9,7 +9,7 @@
 
 <body>
     <div id="container">
-        <canvas width="1200" height="900"></canvas>
+        <canvas width="1200" height="800"></canvas>
         <div id="infoGames">
             <p>Score :
                 <p id="score"></p>
