+++ modules.js/ennemy.js
@@ -1,0 +1,9 @@
+import Entity from './entity';
+
+class Ennemy extends Entity {
+    constructor(posX, posY) {
+        super(posX, posY, 25, 10);
+    }
+}
+
+export default Ennemy;
